chore: move stack allocator
